% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalPD.R
\name{totalpd}
\alias{totalpd}
\title{Funcao retorna valor do dispendio federal em P&D no
ano especificado, em valores correntes.}
\usage{
totalpd(ano)
}
\arguments{
\item{ano}{um numero inteiro}
}
\value{
Um numero ou vetor com valor do dispendio federal em P&D
}
\description{
Se o ano nao for especificado, a funcao
retorna a serie historica completa disponivel
}
